requirements.txt
google-genai
streamlit
pandas
plotly
statsmodels
